refactor(client): share question answer formatting with native consumers
Jesse requested a larger batch of small, reviewable PRs from the native
mobile checkpoint.

Move the pure question-answer formatter and its behavioral tests into the protocol directory, and update the web composer imports in the same change. This makes the existing answer formatting available to native and package consumers.

Web formatting and question behavior remain unchanged. The helper has no React, browser, store or connection dependency. This does not extract higher-level application state.

The 21 formatter tests and canonical web gate pass, including TypeScript and Biome. Independent Luna review is clean.
